requirements_test.txt
pytest
pytest-cov
pytest-homeassistant-custom-component
bandit>=1.7.1
black>=21.12b0
flake8>=4.0.1
pre-commit>=4.0.0
pyflakes>=2.4.0
gitlint>=0.17.0